% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_derivate_range.R
\name{create_derivate_range}
\alias{create_derivate_range}
\title{Recreate a derivative variable based on their range}
\usage{
create_derivate_range(env, varname, co, to_binary = FALSE)
}
\arguments{
\item{env}{The original variable stacks as \code{\link{SpatRaster}} or \code{\link{stars}}.}

\item{varname}{A \code{\link{character}} of the variable name. Needs to be present in
\code{"env"}.}

\item{co}{A set of coefficients obtained via \code{\link[stats:coef]{stats::coef()}} and a
\code{\link[ibis.iSDM:BiodiversityDistribution-class]{ibis.iSDM::BiodiversityDistribution}} object.}

\item{to_binary}{A \code{\link{logical}} flag if the output should be converted to binary
format or left in the original units (Default: \code{FALSE}).}
}
\value{
A \code{\link{SpatRaster}} object containing the predictor range.
}
\description{
The purpose of this function is to create the range from several derivative
variables. The information to do so is taken from the variable name and it
is assumed that those have been created by \code{\link[ibis.iSDM:predictor_derivate]{ibis.iSDM::predictor_derivate()}} function.

This function return the range of values from the original data that fall within
the set of coefficients. Currently only positive coefficients are taken by default.
}
\details{
\itemize{
\item This function really only makes sense for \code{'bin'}, \code{'thresh'} and
\code{'hinge'} transformations.
\item For \code{'hinge'} the combined \code{min} is returned.
}
}
\note{
This is rather an internal function created for a specific use and project. It
might be properly described in an example later.
}
\examples{
\dontrun{
 # Assuming derivates of temperature have been created for a model, this
 # recreates the range over which they apply.
 deriv <- create_derivate_range(env, varname = "Temperature",
  co = coef(fit), to_binary = TRUE)
}

}
\author{
Martin Jung
}
\keyword{utils}
