% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insights_summary.R
\name{insights_summary}
\alias{insights_summary}
\title{Summarize inSiGHTS into an index}
\usage{
insights_summary(obj, toArea = TRUE, fun = "sum", relative = TRUE)

\S4method{insights_summary}{SpatRaster,logical,character,logical}(obj,toArea,fun,relative)

\S4method{insights_summary}{stars,logical,character,logical}(obj,toArea,fun,relative)
}
\arguments{
\item{obj}{A \code{\link{SpatRaster}} or temporal \code{\link{stars}} object describing with the
applied InSiGHTS outputs from \code{insights_fraction}. If the number of layers is greater
than 1, the parameter \code{"relative"} mgiht be applied.}

\item{toArea}{A \code{\link{logical}} flag whether the suitable habitat should be summarized to area (Default: \code{TRUE})?}

\item{fun}{A \code{\link{character}} indicating the summary function to be applied (Default: \code{'sum'}).
Currently supported are \code{'sum'}, \code{'min'}, \code{'max'}, \code{'median'} and \code{'mean'}.}

\item{relative}{A \code{\link{logical}} flag whether a relative index is to be constructed (Default: \code{TRUE}).}
}
\value{
A \code{\link{data.frame}} with area estimates or the respective indicator.
}
\description{
This function handily summarizes the suitable habitat for a given species or
biodiversity feature into an index. If a single timestep (or object with a single layer)
is provided, this function simply summarizes the suitable area.
}
\references{
\itemize{
\item Baisero, Daniele, Piero Visconti, Michela Pacifici, Marta Cimatti, and Carlo Rondinini. "Projected global loss of mammal habitat due to land-use and climate change." One Earth 2, no. 6 (2020): 578-585.
\item Powers, Ryan P., and Walter Jetz. "Global habitat loss and extinction risk of terrestrial vertebrates under future land-use-change scenarios." Nature Climate Change 9, no. 4 (2019): 323-329.
}
}
\author{
Martin Jung
}
