% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{relChange}
\alias{relChange}
\title{Relative change function}
\usage{
relChange(v, fac = 100)
}
\arguments{
\item{v}{A \code{\link{vector}} with numeric values greater than 2.}

\item{fac}{A \code{\link{numeric}} constant multiplier on the resulting metric.}
}
\value{
A \code{\link{numeric}} vector.
}
\description{
This function calculates the relative change of a vector, taking the first value as
a reference value.
}
\examples{
# Example vector
x <- c(20,6,2,1,15,25)
relChange(x)

}
\author{
Martin Jung
}
\keyword{utils}
