% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_clamp.R
\name{st_clamp}
\alias{st_clamp}
\title{Clamp stars to a specific bounds or range}
\usage{
st_clamp(env, lb = -Inf, ub = Inf)
}
\arguments{
\item{env}{The original variable stacks as \code{\link{SpatRaster}} or \code{\link{stars}}.}

\item{lb}{The lower bound for clamping (Default: \code{-Inf}).}

\item{ub}{The lower bound for clamping (Default: \code{Inf}).}
}
\value{
The same as input \code{'env'}.
}
\description{
This small helper function simply clamps the a given \code{\link{SpatRaster}} or \code{\link{stars}} object \
to a provided range. This can help to adjust for example land-use fractions for use
in \code{insights_fraction()}.
}
\examples{
\dontrun{
  # Use
}

}
\author{
Martin Jung
}
\keyword{utils}
